requirements.txt
annotated-doc==0.0.4
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
filelock==3.24.3
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
packaging==26.0
Pygments==2.19.2
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.2.28
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
setuptools==82.0.0
shellingham==1.5.4
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
util==0.0.1
uv==0.10.6

chromadb
groq
PyMuPDF
streamlit
fastapi
sqlalchemy
pydantic
passlib[bcrypt]
PyJWT
python-multipart
pytest

langchain
langchain-community
langchain-groq
langchain-huggingface
langchain-chroma
langchain-experimental

bcrypt